requirements.txt
dj-database-url==0.4.2
Django==2.0
django-split-settings==0.3.0
mysqlclient==1.3.12
pytz==2017.3
